requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
black==26.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
distro==1.9.0
dotenv==0.9.9
faiss-cpu==1.8.0
fuzzysearch==0.8.1
google-auth==2.48.0
google-genai==1.61.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mypy_extensions==1.1.0
numpy<2
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pyttsx3==2.90
pytokens==0.4.1
RapidFuzz==3.14.3
requests==2.32.5
rsa==4.9.1
SpeechRecognition==3.11.0
sentence-transformers==3.0.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
